Re-arrange folder structure for the dart side (#842)
