+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - AudioCache for web
 - Fixing basic features for Android lower than API 23
 - Fixing error after playing music several times with AudioCache
+- Re-organize folder and file structure on the Dart side (project layout)
 
 ## 0.18.3
 - Fix Float vs Double mixup on Swift that prevent non-integer values for volume/playback
