+++ example/lib/generated_plugin_registrant.dart
@@ -4,7 +4,7 @@
 
 // ignore_for_file: lines_longer_than_80_chars
 
-import 'package:audioplayers/audioplayers_web.dart';
+import 'package:audioplayers/web/audioplayers_web.dart';
 
 import 'package:flutter_web_plugins/flutter_web_plugins.dart';
