+++ lib/audioplayers_api.dart
@@ -1,0 +1,4 @@
+export 'src/api/player_mode.dart';
+export 'src/api/player_state.dart';
+export 'src/api/playing_route.dart';
+export 'src/api/release_mode.dart';
