+++ lib/notifications.dart
@@ -1,0 +1,2 @@
+export 'src/notifications/notification_service.dart';
+export 'src/notifications/player_control_command.dart';
